# packages/api-client/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Aldergate Blinds & Shades
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type ProductCategory = typeof ProductCategory[keyof typeof ProductCategory];


export const ProductCategory = {
  roller: 'roller',
  venetian: 'venetian',
  roman: 'roman',
  shutter: 'shutter',
} as const;

export interface Product {
  id: number;
  name: string;
  category: ProductCategory;
  materials: string;
  fabricOptions: string[];
  description: string;
  /** @minItems 1 */
  images: string[];
  createdAt: string;
}

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  category: ProductCategory;
  /** @minLength 1 */
  materials: string;
  fabricOptions: string[];
  description: string;
  /** @minItems 1 */
  images: string[];
}

export interface ProductUpdate {
  /** @minLength 1 */
  name?: string;
  category?: ProductCategory;
  /** @minLength 1 */
  materials?: string;
  fabricOptions?: string[];
  description?: string;
  /** @minItems 1 */
  images?: string[];
}

export interface AdminLoginInput {
  /** @minLength 1 */
  password: string;
}

export interface AdminSession {
  token: string;
}

export interface QuoteItem {
  productId: number;
  productName: string;
  category: ProductCategory;
}

export interface QuoteRequestInput {
  items: QuoteItem[];
  /** @nullable */
  widthCm?: number | null;
  /** @nullable */
  dropCm?: number | null;
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  phone: string;
  /** @minLength 1 */
  email: string;
  /** @minLength 1 */
  postcode: string;
  preferredDate: string;
  /** @minLength 1 */
  preferredTimeWindow: string;
}

export type QuoteRequestStatus = typeof QuoteRequestStatus[keyof typeof QuoteRequestStatus];


export const QuoteRequestStatus = {
  pending: 'pending',
  contacted: 'contacted',
  completed: 'completed',
} as const;

export interface QuoteRequest {
  id: number;
  items: QuoteItem[];
  /** @nullable */
  widthCm: number | null;
  /** @nullable */
  dropCm: number | null;
  name: string;
  phone: string;
  email: string;
  postcode: string;
  preferredDate: string;
  preferredTimeWindow: string;
  status: QuoteRequestStatus;
  createdAt: string;
}

export type QuoteRequestUpdateStatus = typeof QuoteRequestUpdateStatus[keyof typeof QuoteRequestUpdateStatus];


export const QuoteRequestUpdateStatus = {
  pending: 'pending',
  contacted: 'contacted',
  completed: 'completed',
} as const;

export interface QuoteRequestUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  phone?: string;
  /** @minLength 1 */
  email?: string;
  /** @minLength 1 */
  postcode?: string;
  preferredDate?: string;
  /** @minLength 1 */
  preferredTimeWindow?: string;
  /** @nullable */
  widthCm?: number | null;
  /** @nullable */
  dropCm?: number | null;
  status?: QuoteRequestUpdateStatus;
}

export interface QuoteRequestSummary {
  totalPending: number;
  totalThisWeek: number;
  totalItemsRequested: number;
}

export type ListProductsParams = {
category?: ProductCategory;
};