# packages/api-contract/src/generated/types/productUpdate.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Aldergate Blinds & Shades
 * OpenAPI spec version: 0.1.0
 */
import type { ProductCategory } from './productCategory';

export interface ProductUpdate {
  /** @minLength 1 */
  name?: string;
  category?: ProductCategory;
  /** @minLength 1 */
  materials?: string;
  fabricOptions?: string[];
  description?: string;
  /** @minItems 1 */
  images?: string[];
}