# packages/api-contract/src/generated/types/quoteRequestUpdate.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Aldergate Blinds & Shades
 * OpenAPI spec version: 0.1.0
 */
import type { QuoteRequestUpdateStatus } from './quoteRequestUpdateStatus';

export interface QuoteRequestUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  phone?: string;
  /** @minLength 1 */
  email?: string;
  /** @minLength 1 */
  postcode?: string;
  preferredDate?: Date;
  /** @minLength 1 */
  preferredTimeWindow?: string;
  /** @nullable */
  widthCm?: number | null;
  /** @nullable */
  dropCm?: number | null;
  status?: QuoteRequestUpdateStatus;
}